requirements.txt
langchain-groq==0.2.5
langchain-core==0.3.45
langchain-community==0.3.19
python-dotenv==1.0.1
pandas==2.2.3
streamlit==1.34.0